Edit openapi and remove blockService.js
