+++ hedera-mirror-rest/service/recordFileService.js
@@ -25,6 +25,17 @@ const _ = require('lodash');
 const BaseService = require('./baseService');
 const {RecordFile} = require('../model');
 
+const buildWhereSqlStatement = (whereQuery) => {
+  let where = '';
+  const params = [];
+  for (let i = 1; i <= whereQuery.length; i++) {
+    where += `${i === 1 ? 'where' : 'and'} ${whereQuery[i - 1].query} $${i} `;
+    params.push(whereQuery[i - 1].param);
+  }
+
+  return {where, params};
+};
+
 /**
  * RecordFile retrieval business logic
  */
@@ -33,9 +44,9 @@ class RecordFileService extends BaseService {
     super();
   }
 
-  static recordFileBlockDetailsFromTimestampQuery = `select 
+  static recordFileBlockDetailsFromTimestampQuery = `select
     ${RecordFile.CONSENSUS_END}, ${RecordFile.HASH}, ${RecordFile.INDEX}
-    from ${RecordFile.tableName} 
+    from ${RecordFile.tableName}
     where  ${RecordFile.CONSENSUS_END} >= $1
     order by ${RecordFile.CONSENSUS_END} asc
     limit 1`;
@@ -57,6 +68,24 @@ class RecordFileService extends BaseService {
     }
 
     return new RecordFile(rows[0]);
+  }
+
+  async getBlocks(filters) {
+    const {where, params} = buildWhereSqlStatement(filters.whereQuery);
+
+    const query = `
+      select
+        ${RecordFile.COUNT}, ${RecordFile.HASH}, ${RecordFile.NAME}, ${RecordFile.PREV_HASH}, ${RecordFile.BYTES},
+        ${RecordFile.HAPI_VERSION_MAJOR}, ${RecordFile.HAPI_VERSION_MINOR}, ${RecordFile.HAPI_VERSION_PATCH},
+        ${RecordFile.INDEX}, ${RecordFile.CONSENSUS_START}, ${RecordFile.CONSENSUS_END}
+      from ${RecordFile.tableName}
+      ${where}
+      order by ${RecordFile.INDEX} ${filters.order}
+      limit ${filters.limit}
+    `;
+    const rows = await super.getRows(query, params, 'getBlocks');
+
+    return rows.map((recordFile) => new RecordFile(recordFile));
   }
 }
